requirements.txt
playwright
pytest
pytest-playwright
allure-pytest
Faker